Refactor JoinStrings & ReplaceWithMappings into single JavaStringUtils class
