+++ appinventor/components/tests/com/google/appinventor/components/runtime/util/JavaStringUtilsTest.java
@@ -13,9 +13,9 @@ import java.util.Map;
 import static junit.framework.Assert.assertEquals;
 
 /**
- * Test class for the JavaReplaceWithMappings utility.
+ * Test class for JavaStringUtils utility methods.
  */
-public class JavaReplaceWithMappingsTest {
+public class JavaStringUtilsTest {
   /**
    * Test case for replacing a String with no mappings.
    * The String should not be modified.
@@ -25,7 +25,7 @@ public class JavaReplaceWithMappingsTest {
     final String text = "this is a test string";
     Map<Object, Object> mappings = new LinkedHashMap<>();
 
-    final String result = JavaReplaceWithMappings.replaceWithMappings(text, mappings, 1);
+    final String result = JavaStringUtils.replaceWithMappings(text, mappings, 1);
 
     assertEquals(result, text);
   }
@@ -43,7 +43,7 @@ public class JavaReplaceWithMappingsTest {
     mappings.put("tset", "test");
     mappings.put("strxng", "string");
 
-    final String result = JavaReplaceWithMappings.replaceWithMappings(text, mappings, 1);
+    final String result = JavaStringUtils.replaceWithMappings(text, mappings, 1);
 
     assertEquals(result, text);
   }
@@ -60,7 +60,7 @@ public class JavaReplaceWithMappingsTest {
     Map<Object, Object> mappings = new LinkedHashMap<>();
     mappings.put("tset", "test");
 
-    final String result = JavaReplaceWithMappings.replaceWithMappings(text, mappings, 1);
+    final String result = JavaStringUtils.replaceWithMappings(text, mappings, 1);
     final String expected = "this is a test string, testing, testing";
 
     assertEquals(expected, result);
@@ -77,7 +77,7 @@ public class JavaReplaceWithMappingsTest {
     mappings.put("item1", "item2");
     mappings.put("item2", "item3");
 
-    final String result = JavaReplaceWithMappings.replaceWithMappings(text, mappings, 1);
+    final String result = JavaStringUtils.replaceWithMappings(text, mappings, 1);
     final String expected = "Substitute item2 to item3, and then item3 to item3";
 
     // Note how in the expected result, item1 -> item2, rather than
@@ -96,7 +96,7 @@ public class JavaReplaceWithMappingsTest {
     mappings.put(5, 4);
     mappings.put(7,5);
 
-    final String result = JavaReplaceWithMappings.replaceWithMappings(text, mappings, 1);
+    final String result = JavaStringUtils.replaceWithMappings(text, mappings, 1);
     final String expected = "1, 2, 3 and 4, 5, 6";
 
     assertEquals(expected, result);
@@ -112,7 +112,7 @@ public class JavaReplaceWithMappingsTest {
     Map<Object, Object> mappings = new LinkedHashMap<>();
     mappings.put(3.14159f, 3.14f);
 
-    final String result = JavaReplaceWithMappings.replaceWithMappings(text, mappings, 1);
+    final String result = JavaStringUtils.replaceWithMappings(text, mappings, 1);
     final String expected = "Pi: 3.14";
 
     assertEquals(expected, result);
@@ -133,7 +133,7 @@ public class JavaReplaceWithMappingsTest {
     mappings.put("ab", "bc");
     mappings.put("ba", "cb");
 
-    final String result = JavaReplaceWithMappings.replaceWithMappings(text, mappings, 0);
+    final String result = JavaStringUtils.replaceWithMappings(text, mappings, 0);
     final String expected = "bc cb d e";
 
     assertEquals(expected, result);
@@ -154,7 +154,7 @@ public class JavaReplaceWithMappingsTest {
     mappings.put("ab", "bc");
     mappings.put("ba", "cb");
 
-    final String result = JavaReplaceWithMappings.replaceWithMappings(text, mappings, 2);
+    final String result = JavaStringUtils.replaceWithMappings(text, mappings, 2);
     final String expected = "bc ed d e";
 
     assertEquals(expected, result);
@@ -172,7 +172,7 @@ public class JavaReplaceWithMappingsTest {
     mappings.put("a|b", "x");
     mappings.put("c", "y");
 
-    final String result = JavaReplaceWithMappings.replaceWithMappings(text, mappings, 0);
+    final String result = JavaStringUtils.replaceWithMappings(text, mappings, 0);
     final String expected = "a b y d";
 
     assertEquals(expected, result);
